+++ src/main/java/melonslise/locks/common/enchantment/SilentEnchantment.java
@@ -5,9 +5,9 @@ import melonslise.locks.common.init.LocksEnchantments;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.entity.EquipmentSlot;
 
-public class ShockingEnchantment extends Enchantment
+public class SilentEnchantment extends Enchantment
 {
-	public ShockingEnchantment()
+	public SilentEnchantment()
 	{
 		super(Rarity.UNCOMMON, LocksEnchantments.LOCK_TYPE, new EquipmentSlot[] { EquipmentSlot.MAINHAND });
 	}
@@ -15,27 +15,33 @@ public class ShockingEnchantment extends Enchantment
 	@Override
 	public int getMinCost(int level)
 	{
-		return 2 + (level - 1) * 9;
+		return 10;
 	}
 
 	@Override
 	public int getMaxCost(int level)
 	{
-		return this.getMinCost(level) + 30;
+		return 40;
 	}
 
 	@Override
 	public int getMaxLevel()
 	{
-		return 5;
+		return 1;
 	}
 
 	@Override
-	public boolean isDiscoverable() { return LocksServerConfig.ENABLE_SHOCKING.get(); }
+	protected boolean checkCompatibility(Enchantment other)
+	{
+		return super.checkCompatibility(other) && other != LocksEnchantments.SHOCKING.get();
+	}
 
 	@Override
-	public boolean isTradeable() { return LocksServerConfig.ENABLE_SHOCKING.get(); }
+	public boolean isDiscoverable() { return LocksServerConfig.ENABLE_SILENT.get(); }
 
 	@Override
-	public boolean isAllowedOnBooks() { return LocksServerConfig.ENABLE_SHOCKING.get(); }
+	public boolean isTradeable() { return LocksServerConfig.ENABLE_SILENT.get(); }
+
+	@Override
+	public boolean isAllowedOnBooks() { return LocksServerConfig.ENABLE_SILENT.get(); }
 }