+++ src/main/java/melonslise/locks/common/worldgen/LockChestsFeature.java
@@ -7,6 +7,7 @@ import melonslise.locks.common.util.Cuboid6i;
 import melonslise.locks.common.util.ILockableProvider;
 import melonslise.locks.common.util.Lock;
 import melonslise.locks.common.util.Lockable;
+import melonslise.locks.common.util.LootValueCalculator;
 import melonslise.locks.common.util.Transform;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -14,6 +15,8 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.WorldGenLevel;
 import net.minecraft.world.level.block.ChestBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.RandomizableContainerBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.ChestType;
 import net.minecraft.world.level.levelgen.feature.Feature;
@@ -35,11 +38,26 @@ public class LockChestsFeature extends Feature<NoneFeatureConfiguration>
 		RandomSource rng = context.random();
 		BlockPos pos = context.origin();
 
-		if(!LocksConfig.canGen(rng))
-			return false;
+		ItemStack stack;
+		if (LocksConfig.LOOT_SCALED_LOCKS.get())
+		{
+			BlockEntity be = world.getBlockEntity(pos);
+			if (!(be instanceof RandomizableContainerBlockEntity container) || container.lootTable == null)
+				return false;
+			double lootValue = LootValueCalculator.getCachedLootValue(container.lootTable);
+			stack = LocksConfig.getLockForLootValue(lootValue, rng);
+			if (stack == null)
+				return false;
+		}
+		else
+		{
+			if (!LocksConfig.canGen(rng))
+				return false;
+			stack = LocksConfig.getRandomLock(rng);
+		}
+
 		BlockState state = world.getBlockState(pos);
 		BlockPos pos1 = state.getValue(ChestBlock.TYPE) == ChestType.SINGLE || ModList.get().isLoaded("lootr") ? pos : pos.relative(ChestBlock.getConnectedDirection(state));
-		ItemStack stack = LocksConfig.getRandomLock(rng);
 		Lockable lkb = new Lockable(new Cuboid6i(pos, pos1), Lock.from(stack), Transform.fromDirection(state.getValue(ChestBlock.FACING), Direction.NORTH), stack, world.getLevel());
 		lkb.bb.getContainedChunks((x, z) ->
 		{